+++ custom_components/restapi_mypv_p2h/__init__.py
@@ -26,6 +26,7 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
     entry.async_on_unload(entry.add_update_listener(_async_update_options))
+    entry.async_on_unload(coordinator.cancel_keepalive)
     return True